# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SmartLab Backend API
  version: 0.1.0
servers:
  - url: http://47.109.158.254:3000/api/v1
paths:
...
  /auth/pending/{id}/approve:
    post:
      summary: Approve registration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approved
  /auth/pending/{id}/reject:
    post:
      summary: Reject registration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rejected
...